requirements.txt
charset-normalizer==3.4.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pyzbar==0.1.9
qrcode==8.2
reportlab==4.4.4
